[230628] 장바구니 쿠키에 적용

+++ src/main/java/com/winterfoodies/winterfoodies_project/dto/user/CartDto.java
@@ -16,7 +16,7 @@ public class CartDto {
 
     private Long price;
 
-    private int quantity;
+    private Long quantity;
 
     private Long totalPrice;
